+++ src/Cookies/CookiesManager.php
@@ -43,7 +43,7 @@ class CookiesManager implements IteratorAggregate, Countable
     /**
      * @inheritDoc
      */
-    public function getIterator()
+    public function getIterator(): ArrayIterator
     {
         return new ArrayIterator($this->cookies);
     }