+++ CHANGELOG.md
@@ -10,10 +10,11 @@ use [Keep a Changelog] (http://keepachangelog.com/).
 
 - New method `HarParser::addEntryToSession(Session $session, Entry $entry): void`
 - New method `HarParser::getTimings(Entry $entry): Timings`
-- New `callback` option for requests
 - New method `Session::getLastRequest(): ?RequestInterface`
 - New method `Session::getLastResponse(): ?ResponseInterface`
 - New method `Session::createFromHarFile(): Session`
+- New `callback` option for requests
+- New `followLocation` option for requests, return 3xx Response object if it's a redirection
 
 ### Changed