+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [1.3.0] - 2021-11-26
+
+### Added
+
+- PHP 8.1 compatibility
+
 ## [1.2.0] - 2021-07-15
 
 ### Changed